third_party/blink/web_tests/external/wpt/mathml/relations/css-styling/display-2-expected.txt

This is a testharness.js-based test.
[FAIL] block display with column width (math)
  assert_approx_equals: block size expected 78 +/- 1 but got 45
[FAIL] block display with column width (mrow)
  assert_approx_equals: block size expected 78 +/- 1 but got 45
Harness: the test ran to completion.
